refactor: Route 컴포넌트를 별도 폴더로 이동
